feat: Simplify 'Ctrl+S' Functionality in Editor to Prevent Default Save Prompt (#1053)
* feat: Simplify 'Ctrl+S' Functionality in Editor to Prevent Default Save Prompt

- Add a focus handler to set `wrapperFocused` state to `true` and a blur handler to set `wrapperFocused` state to `false`
- Add a save handler that listens for Ctrl + S key combination and displays a toast notification with the title "Saved" and the description "Your code has been saved"

* Update focus state handling in SplitEditor component

- Update variable name from `wrapperFocused` to `isWrapperInnerFocused` to improve clarity and readability.

* Remove state in place of checking the ref in the save handler

The commit removes the state in the SplitEditor component and replaces it with checking the ref in the save handler inside the useEffect hook. This simplifies the code and improves performance.

* Add mac meta key check

The commit adds a check for the meta key (command key) on Mac systems to handle a save event in the Split Editor component. This allows for consistent behavior across different platforms.

---------

Co-authored-by: bautistaaa <chrisbautistaaa@gmail.com>