feat(monaco): respect challenge's tsconfig
This patch also makes `SplitEditor` completely overwrite the TS config on
mount, as it's global. This ensures that it's properly reset for every
challenge.

The default Monaco config doesn't set much: just `allowNonTsExtensions`
and `target`, and we already overwrite the latter. I've added the former
to preserve the defaults.

Reference:
https://github.com/microsoft/monaco-editor/blob/091550e89857bd926d2b66e100dc3c2094475783/src/language/typescript/monaco.contribution.ts#L753