feat(db): override tsconfig for "assert array index"
This patch overrides the tsconfig for challenge 925: Assert Array Index,
which requires `noUncheckedIndexedAccess` to be set.

It is theoretically possible to extract the config from the TS playground
URL that tsch.js.org redirects to for each challenge, but even doing a
HEAD request for every challenge appears to be too slow in practice.
Given that this is the only type challenge that requires special config
options as of today, I don't think it makes sense doing that.