1174 mdwiltfong update validate.ts (#1457)
* Filter out 'aot' challenge from getChallengeIds()

* Add support for AOT challenges in getChallengeIds()

* feat: Update metadata schema for AOT challenges

This commit updates the metadata schema for AOT challenges in the `metadata.schema.json` file. The `maxLength` property for the `label` field has been increased from 7 to 30 characters, and the `prerequisites` field now accepts an array of strings instead of a single string. These changes ensure that the metadata for AOT challenges is properly validated and aligned with the requirements.

---------

Co-authored-by: Jim Burbridge <jhechtf@gmail.com>
Co-authored-by: bautistaaa <chrisbautistaaa@gmail.com>